Vouchers & Multi-community support (#30)
* [bumped] version

* [added] early instantiation of wallet service

* [refactored] all modals into their own folder

* [removed] wallet service in favor of wallet service 2

* [fixed] wrong color around header text

* [updated] iOS 13 target and major dependencies

* [fixed] errors from major deps upgrade

* [fixed] errors from major deps upgrade

* [updated] just set receive qr to zero based on user input

* [fixed] avoid clearing search results with exact match

* [added] show the exact match

* [removed] print

* [added] contacts table to db

* [added] separate account screen

* [added] profile caching

* [updated] contact search to search local db

* [removed] contacts screen temporarily

* [added] edit button again instead of pencil

* [added] settings page as a sub page to account

* [updated] button style and alignment on wallet switcher

* [added] latest web3dart

* [added] new send ux

* [added] sending via camera or recipient

* [added] save profiles as they are found

* [bumped] version

* [added] contacts screen

* [bumped] version

* [fixed] blinking contacts page loading

* [bumped] version

* [fixed] scroll behavior on contacts screen

* [bumped] version

* [added] voucher table

* [added] voucher creation

* [added] voucher creation and sharing

* [updated] voucher text

* [added] blurry background to send slider

* [fixed] issue with missing profile being considered an invalid address

* [added] voucher management for the creator

* [added] archive vouchers to keep a local reference

* [bumped] version

* [fixed] db migration

* [fixed] db migration

* [bumped] version

* [fixed] db migration

* [bumped] version

* [fixed] db migration

* [bumped] version

* [fixed] db migration

* [bumped] version

* [fixed] voucher link length

* [bumped] version

* [removed] unused variable

* [added] voucher viewing

* [bumped] version

* [added] xcode cloud post clone script

* [added] universal link ios

* [added] config service

* [added] default global config file

* Fix update profile for accents and emojis. Better cache with grace

* [updated] app icon

* [updated] app icon

* [updated] app icon

* [fixed] gas calculator

* [added] alias configs

* [bumped] version

* [fixed] only lowercase usernames

* [updated] favicons

* [bumped] version

* [fixed] wallet switching

* [fixed] replay functionality

* [bumped] version

* [removed] unused imports

* Don't print to the console the url, otherwise it gets sent to Sentry

* [fixed] tell the user that the photo is required

* [bumped] version

* [added] voucher deep link handling

* [fixed] voucher redeem loading

* [added] font size configuration on header

* [added] alias and creator to voucher

* [bumped] version

* [updated] voucher viewer

* [added] message for empty voucher

* [fixed] bump nock to fix tests

* [added] ability to set max amount

* [fixed] update voucher cache on read and open

* [bumped] version

* [added] dark mode launch screen

* [fixed] profile circle background pulsing container generating high CPU usage

* [fixed] issue with CPU usage + instantiating wallet logic multiple times

* [bumped] version

* [updated] modules

* [fixed] update voucher state on creation

* [fixed] use proper factory for web to persist

* [bumped] version

* [fixed] make sure we load sqlite3.wasm for web

* [bumped] version

---------

+++ lib/services/db/db.dart
@@ -1,10 +1,11 @@
 import 'dart:io';
 
-import 'package:citizenwallet/services/db/transactions.dart';
+import 'package:citizenwallet/services/db/contacts.dart';
+import 'package:citizenwallet/services/db/vouchers.dart';
 import 'package:flutter/foundation.dart';
 import 'package:path/path.dart';
 import 'package:sqflite_common_ffi_web/sqflite_ffi_web.dart';
-import 'package:sqflite/sqflite.dart';
+import 'package:sqflite_common/sqflite.dart';
 
 abstract class DBTable {
   final Database _db;
@@ -20,7 +21,9 @@ abstract class DBTable {
     )
   ''';
 
-  Future<void> migrate(Database db, int version) async {}
+  Future<void> create(Database db);
+
+  Future<void> migrate(Database db, int oldVersion, int newVersion);
 }
 
 class DBService {
@@ -32,24 +35,44 @@ class DBService {
 
   DBService._internal();
 
-  late Database _db;
-  late TransactionTable transactions;
+  Database? _db;
+  late ContactTable contacts;
+  late VouchersTable vouchers;
 
 // open a database, create tables and migrate data
   Future<Database> openDB(String path) async {
-    final db = await openDatabase(
-      path,
+    final options = OpenDatabaseOptions(
       onConfigure: (db) async {
-        // instantiate a transactions table
-        transactions = TransactionTable(db);
+        // instantiate a contacts table
+        contacts = ContactTable(db);
+
+        // instantiate a vouchers table
+        vouchers = VouchersTable(db);
       },
       onCreate: (db, version) async {
         // migrate data
-        await transactions.migrate(db, version);
+        await contacts.create(db);
+
+        // migrate data
+        await vouchers.create(db);
 
         return;
       },
-      version: 1,
+      onUpgrade: (db, oldVersion, newVersion) async {
+        // migrate data
+        await contacts.migrate(db, oldVersion, newVersion);
+
+        // migrate data
+        await vouchers.migrate(db, oldVersion, newVersion);
+
+        return;
+      },
+      version: 5,
+    );
+
+    final db = await databaseFactory.openDatabase(
+      path,
+      options: options,
     );
 
     return db;
@@ -58,8 +81,21 @@ class DBService {
   Future<void> init(String name) async {
     if (kIsWeb) {
       // Change default factory on the web
-      databaseFactory = databaseFactoryFfiWeb;
+      final swOptions = SqfliteFfiWebOptions(
+        sqlite3WasmUri: Uri.parse('sqlite3.wasm'),
+        indexedDbName: '$name.db',
+      );
+
+      final webContext = await sqfliteFfiWebLoadSqlite3Wasm(swOptions);
+
+      databaseFactory =
+          createDatabaseFactoryFfiWeb(options: webContext.options);
+      // databaseFactory = databaseFactoryFfiWeb;
       // path = 'my_web_web.db';
+    }
+
+    if (_db != null && _db!.isOpen) {
+      _db!.close();
     }
 
     final path =
@@ -69,22 +105,34 @@ class DBService {
 
   // reset db
   Future<void> resetDB() async {
-    final path = _db.path;
-    await _db.close();
+    if (_db == null) {
+      return;
+    }
+
+    final path = _db!.path;
+    await _db!.close();
     await deleteDatabase(path);
     _db = await openDB(path);
   }
 
   // delete db
   Future<void> deleteDB() async {
-    final path = _db.path;
-    await _db.close();
+    if (_db == null) {
+      return;
+    }
+
+    final path = _db!.path;
+    await _db!.close();
     await deleteDatabase(path);
   }
 
   // get db size in bytes
   Future<int> getDBSize() async {
-    final path = _db.path;
+    if (_db == null) {
+      return 0;
+    }
+
+    final path = _db!.path;
     final file = File(path);
     return file.length();
   }
